% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_factor.R
\name{pp_factor}
\alias{pp_factor}
\title{Create a pp module which converts columns to factor.}
\usage{
pp_factor(cols = NULL, desc = NULL, data = NULL, standby = TRUE)
}
\arguments{
\item{cols}{Names of features to convert.}

\item{desc}{(optional) Description of the module.}
}
\description{
Create a pp module which converts columns to factor.
}
